requirements.txt
annoying==0.1.0
asgiref==3.7.2
asttokens==2.0.8
atomicwrites==1.4.1
attrs==22.2.0
autopep8==2.0.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==2.0.12
colorama==0.4.5
debugpy==1.6.3
decorator==5.1.1
Django==4.2.3
django-appconf==1.0.5
django-multiupload==0.6.1
django-phonenumber-field==7.1.0
django-select2==8.1.2
entrypoints==0.4
executing==1.0.0
Faker==12.0.1
flake8==6.0.0
idna==3.4
iniconfig==2.0.0
ipykernel==6.15.3
ipython==8.5.0
jedi==0.18.1
jupyter-core==4.11.1
jupyter_client==7.3.5
matplotlib-inline==0.1.6
mccabe==0.7.0
mixer==7.1.2
mypy==0.991
mypy-extensions==0.4.3
nest-asyncio==1.5.5
packaging==21.3
parso==0.8.3
phonenumbers==8.13.15
pickleshare==0.7.5
Pillow==8.3.1
pluggy==0.13.1
prompt-toolkit==3.0.31
psutil==5.9.2
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.13.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
pytils==0.4.1
pytz==2022.7.1
pywin32==304
pyzmq==24.0.0
requests==2.26.0
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.3
stack-data==0.5.0
toml==0.10.2
tomli==2.0.1
tornado==6.2
traitlets==5.4.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.14
wcwidth==0.2.5
